% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_read_d1_data.R
\name{dl_read_d1_data}
\alias{dl_read_d1_data}
\title{Download and read data along with metadata into your R environment}
\usage{
dl_read_d1_data(data_url, path, fnc = "read_csv", ...)
}
\arguments{
\item{data_url}{(character) An identifier or URL for a DataONE object to download.}

\item{path}{(character) Path to a directory to download data to.}

\item{fnc}{(character) Function to be used to read the data (default is \code{readr::read_csv()}).}

\item{...}{Parameters to pass onto the function specified in \code{fnc}.}
}
\value{
(list) A named list containing data and metadata as data.frames.
}
\description{
This is a convenience function to both download and read data and metadata
into your R environment.
}
\examples{
\dontrun{
dl_read_d1_data("urn:uuid:a2834e3e-f453-4c2b-8343-99477662b570", path = "./Data")
}
}
\seealso{
\code{\link[=download_d1_data]{download_d1_data()}} \code{\link[=read_d1_files]{read_d1_files()}}
}
