% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_d1_data_pkg.R
\name{download_d1_data_pkg}
\alias{download_d1_data_pkg}
\title{Download all data and metadata of a data package from DataONE}
\usage{
download_d1_data_pkg(meta_obj, path)
}
\arguments{
\item{meta_obj}{(character) A DOI or metadata object PID for a DataONE package to download.}

\item{path}{(character) Path to a directory to download data to.}
}
\value{
(list) Paths where data are downloaded to.
}
\description{
Downloads all the data objects of a data package from DataONE along with metadata.
}
\examples{
\donttest{
download_d1_data_pkg("doi:10.18739/A2CJ87M3J", tempdir())
download_d1_data_pkg("https://doi.org/10.18739/A2CJ87M3J", tempdir())
}
}
\seealso{
[read_d1_files()] [download_d1_data()]
}
