% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_D1Data.R
\name{download_D1Data}
\alias{download_D1Data}
\title{Downloads data from DataOne along with metadata}
\usage{
download_D1Data(data_obj, path)
}
\arguments{
\item{data_obj}{(character) An identifier or url for a DataONE object to download.
If the object is a metadata or resource map object, all data associated with that object will downloaded.}

\item{path}{(character) Path to a directory to download data to}
}
\value{
(character) Path where data is downloaded to
}
\description{
Downloads data from DataOne along with metadata
}
