% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ISO_data.R
\name{download_ISO_data}
\alias{download_ISO_data}
\title{Download data and metadata from a dataset that uses ISO metadata.}
\usage{
download_ISO_data(meta_raw, meta_obj, meta_id, data_id, metadata_nodes, path)
}
\arguments{
\item{meta_raw}{(character) A raw metadata object produced by download_d1_data}

\item{meta_obj}{(character) A metadata object produced by download_d1_data}

\item{meta_id}{(character) A metadata identifier produced by download_d1_data}

\item{data_id}{(character) A data identifier produced by download_d1_data}

\item{metadata_nodes}{(character) The member nodes where this metadata is stored, produced by download_d1_data}

\item{path}{(character) Path to a directory to download data to.}
}
\description{
This is an internal function called by the download_d1_data.R function. Not to be exported
}
\keyword{internal}
