% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_pids.R
\name{get_pkg_pids}
\alias{get_pkg_pids}
\title{Get pids from a package DOI}
\usage{
get_pkg_pids(pkg_doi)
}
\arguments{
\item{pkg_doi}{(character) The package DOI, or the metadata object PID on a DataONE member node #TODO: figure out terminology}
}
\description{
This function takes a package DOI and optionally, a data URL and returns the associated pids. Alternative to \code{url_to_pid()}, based on \code{arcticdatautils::get_package()}.
}
\examples{
\dontrun{
#get all pids in the data package
get_pkg_pids("https://doi.org/10.6073/pasta/cd7ec5009430eb78f464a66f6794579f")
}

}
