% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metajam-package.R
\docType{package}
\name{metajam}
\alias{metajam}
\title{\code{metajam} package}
\description{
Metadata jam - bringing data and metadata together
}
\details{
See \href{https://github.com/nceas/metajam}{GitHub}
}
