% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_d1_files.R
\name{read_d1_files}
\alias{read_d1_files}
\title{Read data along with metadata based on dowload_d1_data file structure}
\usage{
read_d1_files(folder_path, fnc = "read_csv")
}
\arguments{
\item{folder_path}{(character) Path to a directory where data and metadata are located}

\item{fnc}{(character) Function to be used to read the data (default is readr::read_csv)}
}
\value{
(list) named list containing data and metadata as data frames
}
\description{
Read data along with metadata based on dowload_d1_data file structure
}
\examples{
\dontrun{
data_folder <- system.file("extdata", "doi_10.18739_A23F4KM7K__Alexander_Exp Burn Soil Mois 2012_2017", package = "metajam")
read_d1_files(data_folder)
}

}
