% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_d1_data.R
\name{download_d1_data}
\alias{download_d1_data}
\title{Downloads data from DataOne along with metadata}
\usage{
download_d1_data(data_obj, path)
}
\arguments{
\item{data_obj}{(character) An identifier or url for a DataONE object to download.}

\item{path}{(character) Path to a directory to download data to}
}
\value{
(character) Path where data is downloaded to
}
\description{
Downloads data from DataOne along with metadata
}
