% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_d1_data_pkg.R
\name{download_d1_data_pkg}
\alias{download_d1_data_pkg}
\title{Download all the datasets from a data package}
\usage{
download_d1_data_pkg(meta_obj, path)
}
\arguments{
\item{meta_obj}{(character) A DOI for a DataONE package to download.}

\item{path}{(character) Path to a directory to download data to}
}
\value{
(character) Path where data is downloaded to
}
\description{
Download all the datasets from a data package
}
\examples{
\dontrun{
download_d1_data_pkg("doi:10.18739/A2028W")
download_d1_data_pkg("https://doi.org/10.18739/A2028W")
}

}
